+++ infra/terraform/1_network.tf
@@ -1,14 +1,15 @@
 module "network" {
-  source = "git::https://github.com/Bh-an/sc-tf-service-host-module.git//terraform/modules/network?ref=v0.3.4"
+  source = "git::https://github.com/Bh-an/sc-tf-service-host-module.git//terraform/modules/network?ref=v0.3.5"
 
   region               = var.region
   platform             = var.platform
   environment          = var.environment
   vpc_cidr             = var.vpc_cidr
   public_subnet_cidrs  = var.public_subnet_cidrs
-  private_subnet_cidrs = []
+  private_subnet_cidrs = var.private_subnet_cidrs
   db_subnet_cidrs      = []
   availability_zones   = var.availability_zones
   single_nat_gateway   = true
+  enable_nat_gateways  = var.enable_nat_gateways
   eks_cluster_name     = null
 }